Simplified codebase by merging similar files
All unit and unit type definitions are now located in the `unit.go` file.
A function for generic conversion and functions calling this are implemented in the `convert.go` file. This means this refactor will not be breaking the suggested api. Test functions are also compiled into a single test file `convert_test.go`.

``` go
...
uc.Convert(<value>, <from>, <to>) # generic convert function
uc.Length(<value>, <from>, <to>) # will check first if unit types are Length Units

```
