+++ convert.go
@@ -26,3 +26,37 @@ func Convert(v float64, from, to Unit) (float64, error) {
 
 	return v * factor, nil
 }
+
+//
+
+func Length(v float64, from, to Unit) (float64, error) {
+	if from.Type != LengthUnit || to.Type != LengthUnit {
+		return 0, invalidInputTypes
+	}
+
+	return Convert(v, from, to)
+}
+
+func Area(v float64, from, to Unit) (float64, error) {
+	if from.Type != AreaUnit || to.Type != AreaUnit {
+		return 0, invalidInputTypes
+	}
+
+	return Convert(v, from, to)
+}
+
+func Velocity(v float64, from, to Unit) (float64, error) {
+	if from.Type != VelocityUnit || to.Type != VelocityUnit {
+		return 0, invalidInputTypes
+	}
+
+	return Convert(v, from, to)
+}
+
+func Discharge(v float64, from, to Unit) (float64, error) {
+	if from.Type != DischargeUnit || to.Type != DischargeUnit {
+		return 0, invalidInputTypes
+	}
+
+	return Convert(v, from, to)
+}
