--- units.go
@@ -1,12 +1,0 @@
-package unitconvert
-
-type UnitType int
-
-// Constants representing different types of measurement units.
-const (
-	LengthUnit UnitType = iota
-	AreaUnit
-	VolumeUnit
-	VelocityUnit
-	DischargeUnit
-)
